+++ src/Pia.Wpf/Views/ResearchHistoryView.xaml.cs
@@ -8,4 +8,12 @@ public partial class ResearchHistoryView : UserControl
     {
         InitializeComponent();
     }
+
+    private void OnEntriesSelectionChanged(object sender, SelectionChangedEventArgs e)
+    {
+        if (sender is ListBox listBox && listBox.SelectedItem is { } selected)
+        {
+            listBox.ScrollIntoView(selected);
+        }
+    }
 }